requirements.txt
cachetools>=5.3.1
celery>=5.3.5
celery-types>=0.11.0
colorlog>=6.8.0
Django>=3.2.17,<3.2.25
djangorestframework>=3.15.0
psycopg2>=2.9.8
drf-spectacular>=0.21.2
django-axes[ipware]>=5.30.0
django-cors-headers>=3.11.0
django-silk>=4.3.0
django-split-settings>=1.0.0
django-storages>=1.13
django-redis>=5.2.0
django-celery-results>=2.3.0
django-celery-results-rest>=0.0.1
django-celery-beat>=2.3.0
djangorestframework-simplejwt>=5.1.0
qrcode>=7.3.1
django-allauth>=0.48.0
webauthn>=1.2.1
PyJWT>=2.4.0
python-dotenv>=0.20.0
python3-saml>=1.13.0
requests>=2.27.1
requests-oauthlib>=1.3.1
python3-openid>=3.2.0
xmlsec>=1.3.13
lxml>=4.8.0
dj-rest-auth>=2.2.2
pandas>=1.4.0
openpyxl>=3.0.10
scikit-learn>=1.1.0
xlrd>=2.0.1
pytz>=2023.4
redis>=4.1.1
sentry-sdk>=1.5.2
gunicorn>=21.0.0
typing-extensions>=4.1.0
boto3>=1.34.129
dj_database_url>=2.1.0
SQLAlchemy>=1.4.30
sqlalchemy-celery-beat>=0.8.0
asgiref>=3.5.0
reportlab>=2.0
spacy>=2.3.3,<3.1.0
numpy==1.26.4
python-dotenv